move crypto code to separate object file
