+++ src/nts_extfields.h
@@ -3,13 +3,12 @@
 #include "nts.h"
 
 struct NTS_query {
-	struct NTS_cookie cookie;
-	const uint8_t *c2s_key, *s2c_key;
 #ifndef USE_LIBAES_SIV
 	EVP_CIPHER *cipher;
-#else
-	unsigned int key_len;
 #endif
+	struct NTS_cookie cookie;
+	const uint8_t *c2s_key, *s2c_key;
+	unsigned int key_len;
 };
 
 struct NTS_receipt {
