+++ src/nts_fuzz.c
@@ -10,46 +10,50 @@
 /* this program does no sanity checking as it is meant for fuzzing only */
 
 int main(int argc, char **argv) {
-        int file = open(argv[1], O_RDONLY);
-        unsigned char buffer[1280];
-        int len = read(file, buffer, 1280);
-        if(len < 48) return 0;
-        int aead_id = NTS_AEAD_AES_SIV_CMAC_256;
+	int file = open(argv[1], O_RDONLY);
+	unsigned char buffer[1280];
+	int len = read(file, buffer, 1280);
+	if(len < 48) return 0;
+	int aead_id = NTS_AEAD_AES_SIV_CMAC_256;
 
-        struct NTS_query nts = (struct NTS_query) {
+	struct NTS_query nts = (struct NTS_query) {
 #ifndef USE_LIBAES_SIV
-                .cipher = NTS_AEAD_cipher(aead_id),
-#else
-                .key_len = NTS_AEAD_key_size(aead_id),
+		.cipher = NULL,
 #endif
-                .c2s_key = (void*)"01234567890abcdef",
-                .s2c_key = (void*)"01234567890abcdef",
-        };
+		.key_len = NTS_AEAD_key_size(aead_id),
+		.c2s_key = (void*)"01234567890abcdef",
+		.s2c_key = (void*)"01234567890abcdef",
+	};
 
-        if(argc > 2) {
-                /* fuzz the nts ke */
-                struct NTS_agreement rec;
-                (void) NTS_decode_response(buffer, len, &rec);
-        } else {
-                struct NTS_receipt rcpt = { 0, };
-                (void) NTS_parse_extension_fields(&buffer, len, &nts, &rcpt);
-        }
+	if(argc > 2) {
+		/* fuzz the nts ke */
+		struct NTS_agreement rec;
+		(void) NTS_decode_response(buffer, len, &rec);
+	} else {
+		struct NTS_receipt rcpt = { 0, };
+		(void) NTS_parse_extension_fields(&buffer, len, &nts, &rcpt);
+	}
 
-        return 0;
+	return 0;
 }
 
-/* make sure encryption is a no-op --- currently only when LIBAES_SIV is used */
+/* null cipher */
 
-#define STUB(name) int name() { return 1; }
+#define BLKSIZ 16
 
-#ifndef USE_LIBAES_SIV
-#  error can only fuzz when configured for libaes_siv, currently
-#else
-/* stub libaes_siv functions */
-STUB(AES_SIV_Init)
-STUB(AES_SIV_AssociateData)
-STUB(AES_SIV_EncryptFinal)
-STUB(AES_SIV_DecryptFinal)
-STUB(AES_SIV_CTX_free)
-void *AES_SIV_CTX_new() { return ""; }
-#endif
+int NTS_encrypt(unsigned char *ctxt, const unsigned char *ptxt, int ptxt_len, const void *info, const void *nts) {
+	(void) info;
+	(void) nts;
+	memset(ctxt, 0xEE, BLKSIZ);
+	memmove(ctxt+BLKSIZ, ptxt, ptxt_len);
+	return ptxt_len + BLKSIZ;
+}
+
+int NTS_decrypt(unsigned char *ptxt, const unsigned char *ctxt, int ctxt_len, const void *info, const void *nts) {
+	(void) info;
+	(void) nts;
+	if(ctxt_len < BLKSIZ) return -1;
+
+	memmove(ptxt, ctxt+16, ctxt_len - BLKSIZ);
+	return ctxt_len - BLKSIZ;
+}
